пиу-пиу
научил врагов стрелять и вращаться. Снаряды через некоторое время уничтожаются, враги с нулём ХП тоже.

+++ OOPTaskProject/Assets/Scripts/ProjectileData.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using OOPTaskOne;
 
 
 public class ProjectileData : MonoBehaviour
@@ -10,7 +9,7 @@ public class ProjectileData : MonoBehaviour
     //и содержит следующую информацию: тип снаряда, урон, скорость перемещения и время жизни снарядов
     public Missile Missile;
 
-    [Range(10f, 100f)] public float damage;
+    [Range(10f, 100f)] public int damage;
     [Range(10f, 100f)] public float movementSpeed;
     [Range(0.5f, 20f)] public float remainingTime;
 }
